mentionFactory makes the mentionIterators
