+++ core/src/main/scala/ai/lum/odinson/MentionFactory.scala
@@ -1,7 +1,11 @@
 package ai.lum.odinson
 
-import ai.lum.odinson.state.ResultItem
+import ai.lum.odinson.lucene.OdinResults
+import ai.lum.odinson.lucene.search.OdinsonScoreDoc
 import ai.lum.odinson.utils.MostRecentlyUsed
+import com.typesafe.config.Config
+
+import scala.annotation.tailrec
 
 trait IdGetter {
   def getDocId: String
@@ -9,6 +13,43 @@ trait IdGetter {
 }
 
 trait MentionFactory {
+
+  class OdinMentionsIterator(labelOpt: Option[String], nameOpt: Option[String], odinResults: OdinResults, factory: MentionFactory, mruIdGetter: MostRecentlyUsed[Int, LazyIdGetter]) extends Iterator[Mention] {
+    val scoreDocs: Array[OdinsonScoreDoc] = odinResults.scoreDocs
+    val matchesTotal: Int = scoreDocs.foldLeft(0) { case (total, scoreDoc) => total + scoreDoc.matches.length }
+
+    var matchesRemaining: Int = matchesTotal
+    var scoreDocsIndex: Int = 0
+    var matchesIndex: Int = 0
+
+    override def hasNext: Boolean = 0 < matchesRemaining
+
+    @tailrec
+    override final def next(): Mention = {
+      val scoreDoc = scoreDocs(scoreDocsIndex)
+      val docIndex = scoreDoc.doc
+
+      if (matchesIndex < scoreDoc.matches.length) {
+        val odinsonMatch = scoreDoc.matches(matchesIndex)
+
+        matchesIndex += 1
+        matchesRemaining -= 1
+
+        // TODO: double check that name == foundBy, replace?
+        val idGetter = mruIdGetter.getOrNew(docIndex)
+        factory.newMention(odinsonMatch, labelOpt, docIndex, scoreDoc.segmentDocId, scoreDoc.segmentDocBase, idGetter, nameOpt.getOrElse(""))
+      }
+      else {
+        scoreDocsIndex += 1
+        matchesIndex = 0
+        next()
+      }
+    }
+  }
+
+  def mentionsIterator(labelOpt: Option[String], nameOpt: Option[String], odinResults: OdinResults, mruIdGetter: MostRecentlyUsed[Int, LazyIdGetter]): OdinMentionsIterator = {
+    new OdinMentionsIterator(labelOpt, nameOpt, odinResults, this, mruIdGetter)
+  }
 
   // If you are implementing a custom MentionFactory, this is the primary method
   // you need to implement.
@@ -36,16 +77,6 @@ trait MentionFactory {
     newMention(odinsonMatch, label, luceneDocId, luceneSegmentDocId, luceneSegmentDocBase, idGetter, foundBy, arguments)
   }
 
-  def newMention(resultItem: ResultItem, mruIdGetter: MostRecentlyUsed[Int, LazyIdGetter]): Mention = {
-    newMention(
-      resultItem.odinsonMatch,
-      Some(resultItem.label),
-      resultItem.docIndex,
-      resultItem.segmentDocId,
-      resultItem.segmentDocBase,
-      mruIdGetter.getOrNew(resultItem.docIndex),
-      resultItem.name)
-  }
 
   /** A map from argument name to a sequence of matches.
     *
@@ -84,4 +115,8 @@ trait MentionFactory {
       }
   }
 
+}
+
+object MentionFactory {
+  def fromConfig(config: Config): MentionFactory = new DefaultMentionFactory
 }
